reef: S5 and G3 powerdown
For proto we had to leave everything on since shutting off power
would cause I2C communication with the PD tasks to fail. Now the
PD chips are always on and we can shut off power rails and the PMIC
and also put the Analogix chip in standby mode in G3.

BUG=chrome-os-partner:53646
BRANCH=none
TEST=apshutdown on evt does not cause a bunch of errors

Change-Id: Ic8c6497a33016a20fe41524df7a6af0006e612f7
Signed-off-by: David Hendricks <dhendrix@chromium.org>