g: usb_updater: add command line option to allow updating the RO section
This just adds the new command line option and a help message for it.

This option will be used to signal that the RO section update is
required along with the RW section update when invoking usb_updater.

The flag indicating presence of the option will be saved in the
endpoint structure so it is available at different layers of the code.

BRANCH=none
BUG=chrome-os-partner:55789
TEST=built usb_updater and invoked it with --help option, observed the
     new option showing up in the output.

Change-Id: Ief35a553651d29ca39c42c5af7294d3aba223b02
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>