g: use timels to resume from sleep for scheduled events
The high speed clock does not run when cr50 is in sleep. This change
adds a low speed timer and uses it to update the hw timer at resume. It
also sets an alarm so the system will wake up when the next event is
supposed to be triggered.

BUG=chromium:635620
BRANCH=none
TEST=manual
	reduce the sleep delays in board/cr50/board.c and chip/g/idle.c
	so cr50 will go to sleep more quickly.

	add a function that prints something every second. Verify it
	prints every second and use power measurements to cr50 check
	that cr50 is asleep.

	check that cr50 can go into deep sleep and that the print
	statement wont wake it up.

Change-Id: I96b445664e3ca3cb5a82861e5f4dfe4c9e1e92ef