chip/npcx/lpc: Set SCI bit in EC_SC on SCI generation
When SCI needs to be generated, two things are required:
1. Set SCIE bit in HIPMIE register so that EC_SCI event is generated by
PM channels module
2. Set SCIE bit in HIPMST register so that when kernel queries
EC_SC(0x66) it knows that SCI triggered IRQ to the host.

Current code was handling #1 above, however #2 was missing. So, lidclose
did not result in the correct behavior on kernel's side. With this
change, lidclose triggers the correct state sequence: S0->S3->S5->G3.

BUG=chrome-os-partner:55489
BRANCH=None
TEST=lidclose results in state transition S0->S3->S5->G3.

Change-Id: Ib6e0ee32bc06a8c51297010ac4843a7de45a99bd
Signed-off-by: Furquan Shaikh <furquan@google.com>