Update patch set 1
Patch Set 1:

i2c_port_to_controller() actually can return -1 if things are misconfigured, which is why the check was behind the ifdef. Yes, it's true that the port param can initially be '-1', but we aren't checking for it in our other i2c functions. Should we consistent and either validate the port param in all of our non-static i2c functions or none of them?

Patch-set: 1
Label: Code-Review=0