Change transfer_endpoint to transfer_descriptor
In preparation to cr50 update protocol enhancements allowing to update
both RO and RW sections this patch is changing the name of the
structure containing communications endpoint information.

In the upcoming patches more fields are going to be added to this
structure, it is cleaner to change its name and existing use before
adding new features.

BRANCH=none
BUG=chrome-os-partner:55789
TEST=make -C extra/usb_updater

Change-Id: I5ca0b6da75067a4f01ea80b6a6ca49a50e1136ab
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>