btle: Added SN/NESN flow control
Note that there is an @TODO present. Flow control has been added,
but actual data is not being read from or loaded to the packet.
This will come in the next CL.

TEST=printouts comparing the sent and received SNs and NESNs against
what actually is supposed to come as defined by the manual. Not included
in CL.
BUG=None
BRANCH=None

CQ-DEPEND=CL:370597,CL:370839

Change-Id: Ic0abe722e85ad5ec8e308801b6e6d1afb410e194
Signed-off-by: Levi Oliver <levio@google.com>