cr50: enable deep sleep on chipset shutdown
Cr50 monitors UART1 RX to sense the state of the AP. In this change
if cr50 sees that it is not pulled up by the AP then it sets the state
as off and then notifies HOOK_CHIPSET_SHUTDOWN.

On CHIPSET_SHUTDOWN cr50 will set the idle_action to deep sleep. If
sleep is enabled it will go into deep sleep. If not it will wait until
it is sleep is enabled.

This change also sets the idle_action to IDLE_DEFAULT when resuming from
deep sleep or during init. Before cr50 kept track of the previous idle
state in a PWRDN register and then used that state during the next
resume. If we went into deep sleep, on resume we want the idle action to
be reset to sleep and then only enter deep sleep if we have detected the
AP is off.

BUG=chrome-os-partner:56100
BUG=chrome-os-partner:55747
BRANCH=none
TEST=manual
	run 'poweroff' on the AP and see that cr50 enables deep sleep

	verify that even if the ap is powered off it doesn't prevent ccd
	from working and when suzyq is unplugged cr50 will go into deep
	sleep

	After running poweroff on the AP wait a while and run powerbtn
	on the EC. Verify the system can boot up fully without going
	into recovery.

	Do this on gru and reef.

Change-Id: I07f5a9d85dd0467cd22e499d4261c75caf653563
Signed-off-by: Mary Ruthven <mruthven@chromium.org>