cr50: notify chipset hooks when the AP state changes
Cr50 monitors UART1 RX to sense the state of the AP. In this change
if cr50 sees that UART1 RX is not pulled up by the AP then it sets the
state as off and then notifies HOOK_CHIPSET_SHUTDOWN. When UART1 RX is
high again the the AP is back on and HOOK_CHIPSET_RESUME is notified.

Currently nothing is done with these hooks.

BUG=chrome-os-partner:55747
BRANCH=none
TEST=buildall

Change-Id: I47fb38a4bbcd72424ec2535d61e87f820cf1bcd7
Signed-off-by: Mary Ruthven <mruthven@chromium.org>