TEST-ONLY: HACK: Wake back up right away
Normally for suspend/resume testing we set a bit of randomness for how
long we want to suspend for.  That's good for testing all the corner
cases (especially for testing wakups that happen mid-suspend).
However,sometimes you just want to plow through as many suspend/resume
cycles as you can to try to debug problems.  In this case, you really
want to schedule a wakeup immediately after you go to sleep.

This hacky change does just that.  If you pick it into your EC code it
will always schedule an RTC wake right awya.

It's intended that you might use:
  suspend_stress_test \
    --wake_min=0 --wake_max=1 \
    --suspend_min=20 --suspend_max=21

...we set a suspend time of 20 seconds, but that will really be
overridden by this patch.  Having a short wake time allows you to plow
through as many cycles as you can

BUG=None
TEST=None

Change-Id: I4b5405558fb0856703002937c357e83511bfae85
Signed-off-by: Douglas Anderson <dianders@chromium.org>