Update patch set 1
Patch Set 1:

(Low priority) it would be nice to add an EC command for this mode, or maybe add a hack where an RTC alarm of 0xffffffff worked this way?

Anyway, for now I'm just going to pick this when I need fast suspend/resume testing to reproduce bugs.

Patch-set: 1