anx74xx: Add a retry loop in init function
This adds a retry loop in the init function that will attempt to
read the vendor ID bytes. This is similar to what we do in
tcpci_tcpm_init(), since the TCPC init functions might be called
before the EC's I2C module is initialized and before the TCPC
itself is fully booted.

BUG=none
BRANCH=none
TEST=see "TCPC p0 init ready" instead of error on serial console

Change-Id: Id43ff820494a71702c5a6ffcf636d52267b68141
Signed-off-by: David Hendricks <dhendrix@chromium.org>