cr50: ignore sys_rst when ec_rst is asserted
Cr50 may be used to hold the EC in reset for development purposes by 
writing to the RBOX_ASSERT_EC_RST register or while flashing the EC. 
When cr50 holds the EC in reset sys_rst will be assserted. Cr59 should 
not reset itself because of sys_rst when it is asserting EC reset.

BUG=none
BRANCH=none
TEST='rw 0x40550010 1' causes the EC to reset but not cr50

Change-Id: I97b0c3bde895028ae23b9c02212b0b39f02533da