Allow board-specific defaults to "reboot" command
In most boards we distinguish between a "hard" and a "soft"
reboot of the EC, with the default for the console command being
a "soft" reboot. This CL allows individual boards to override
that default case.

BUG=chrome-os-partner:55948
BRANCH=none
TEST=make buildall; test on Cr50 hardware

The Cr50's console command now defaults to a hard reboot with no
args. All other boards are unaffected.

Change-Id: I957f079204a4d79f4a79e7a35a4dfb71802f0fcb
Signed-off-by: Bill Richardson <wfrichar@chromium.org>