Update patch set 2
Patch Set 2: Code-Review+1

Should be enough to capture the problems we saw, so +1.
But, while we are at it, would also be good to check how it handles unaligned starting addresses for out and in. 
E.g. run encryption/decryption twice: on the first run, copy cmd to the local buf with even starting address, on the 2nd run - with odd starting address, and compare that the results of both runs are the same.

Patch-set: 2
Label: Code-Review=+1