Update patch set 3
Patch Set 3:

as I said, enough for now, so +1. new code for me, so don't dare to +2.
in fact, as I wrote, in the future worth checking both aligned and non-aligned (internal logic may treat unaligned differently and use slower routines or intermediate buffers, but aligned then are also worth checking separately) + in case of unaligned worth checking that byte before the beg of the buffer, if it was overwritten. but these are minor points.

Patch-set: 3