Update patch set 2
Patch Set 2:

> If it ever happens again and you have an EC console open, try to
 > grab the console log and the output of 'gpioget'.

Sure.  My current theory is that it's related to the hack CL I'm using for testing, CL:373104.

In a working case I saw:

 [902.670972 chipset -> S3]
 [902.671192 WBU: Setting alarm]
 [902.739288 power state 2 = S3, in 0x0003]
 [902.739654 power state 2 = S3, in 0x0003]
 [903.289945 event set 0x02000000]
 [903.290781 power state 2 = S3, in 0x000b]

In a failing case I saw:

 [912.310910 chipset -> S3]
 [912.311130 WBU: Setting alarm]
 [912.313359 event set 0x02000000]
 [912.379106 power state 2 = S3, in 0x000b]

I think that the failure may happen if we we get the wakeup event at just the wrong time.  This could be an rk3399 problem and could be related to http://crosbug.com/p/56545 ?

For now I'm going to go back to testing with your patch but going to change my HACK CL to try to wake up in 1 to 2 seconds (instead of the current 0 to 1).  Once we get normal suspend/resume happy we can worry about testing suspend/resume with poorly timed wakeup events.

Patch-set: 2
Label: Code-Review=0