kevin: Turn on pull down on SPI chip select
The SPI lines are an input to the EC from the AP and the AP might not be
turned on.  We should have a pull down on these lines to avoid them
glitching when the AP is in S3 or S5.

I'm not sure how to turn a pulldown on the main SPI lines, but the chip
select is easy and the most important.  Do that here.

BRANCH=None
BUG=chrome-os-partner:56644
TEST=Helps hl@

Change-Id: I3a9b018e6e8a5eddb1f23e004f1af3da3e503709
Signed-off-by: Douglas Anderson <dianders@chromium.org>