Cr50: Enable restricted console commands
This adds a "lock" command to enable or disable the restricted
console commands. For now, the lock is disabled so all commands
continue to work as before, and even after enabling it, it's
trivial to disable it again. Future CLs will build on this
framework.

BUG=chrome-os-partner:55322
BRANCH=none
TEST=make buildall, test on Cr50 hardware

Try these commands:

  lock
  help
  gpioget
  lock enable
  help
  gpioget
  lock disable
  help
  gpioget

Change-Id: I42c9bd63e17612dcff78c9f45054e53d96adcd5b
Signed-off-by: Bill Richardson <wfrichar@chromium.org>