Update patch set 6
Patch Set 6:

(1 comment)

sorry, this is no the list of things to try real soon now.

One extra complication with this is that we don't have a way to restart the H1 after update once this lands other than powercyling the device.

We could/should come up with an H1 command to do it, but this patch should not land before that mechanism is in place.

Patch-set: 6