Update patch set 4
Patch Set 4:

I've rebased this so it will merge cleanly on top of some other CLs that are about to go in, and added a couple of reviewers since it's related to a previously reported bug.

I didn't see any problems on my Kevin, but it would be nice to confirm on other boards too. If anyone has some, please chime in here.

Patch-set: 4