Update patch set 2
Patch Set 2:

Hrm.  I wonder if it makes sense to set the chip select line as "pull up" when the AP is on and "pull down" when the AP is off.  Is that possible?

Maybe not important, but I'm worried that we'll see a chip select assertion when the AP goes into S3...

Patch-set: 2