Cr50: Add gpio input for platform reset (plt_rst_l)
For TPM operation with Intel chipset APs, the signal PLT_RST_L needs
to trigger a TPM reset. For Reef, starting with DVT, this signal is
connected to DIOM3 on H1. For Reef EVT and existing Gru/Kevin DIOM3 is
connected to AP_FLASH_WP_L. This line will not transition during
normal operation so should not cause TPM resets with existing boards.

BRANCH=none
BUG=chrome-os-partner:55115
TEST=manual
Used H1 dev board and connected a GPIO output to DIOM3 pin. Created
high to low transition to verify that sys_reset was detected. Tested
on Kevin/Reef to verify no resets were not occurring.

Change-Id: I58f02b7ffa644a9197f4303ae6e640df181040bd
Signed-off-by: Scott <scollyer@chromium.org>