pd: Add individual chip granularity for PD reset function
This is intended to allow us to reset individual TCPC chips. This
is mostly to support ectool commands for development, but could
also be useful for recovering from errors with minimal disruption.
For example, if somebody is copying their precious baby pictures from
USB flash connected on one port but then plug in a non-Benson approved
charger into a port controlled by a different TCPC which subsequently
goes haywire and needs to be reset.

Also, it can help when dealing with Analogix TCPCs since they tend
to require huge delays when resetting which may cause issues with
other TCPCs in the system (see Kevin's comment in CL:347067).

BUG=none
BRANCH=none
TEST=make buildall -j

Change-Id: I457ff60fffdf32113d7b8b6a69a740f0c48408be
Signed-off-by: David Hendricks <dhendrix@chromium.org>