Update patch set 4
Patch Set 4:

(8 comments)

This is higher up in the stack, so I don't have too much insight, but it seems that today Chrome is the one that does tablet mode detection. However, it looks like we're moving that to the EC. Do the angles defined here match what already existing behavior?

Patch-set: 4
Label: Code-Review=0