COMMIT-QUEUE.ini: Change builders to compile-only.
(Re-attempt of https://chromium-review.googlesource.com/#/c/372325/)
Really, we only care if the chromeos-ec package fails to build and not
if the test stages fail.  The test stages don't actually test aganist
our EC changes anways, so it's kind of a waste of time to run them.
Besides, that's what FAFT is for.

BUG=chromium:642503
BRANCH=None
TEST=cbuildbot --remote chell-compile-only-pre-cq

CQ-DEPEND=CL:381323

Change-Id: I1b4b7fc68a9f8a943f6f5ef3d8b169264c95359e
Signed-off-by: Aseda Aboagye <aaboagye@google.com>