Update patch set 6
Patch Set 6: Code-Review-2

Hi Vijay,
Upon revisiting this, I noticed that section 2 of the ANX74xx "Programming Guide" has a slightly different explanation for RESETN and PWR_EN than I had originally gleaned from section 5 of the "Datasheet." Here's a table to summarize:
          Reset  |  Standby  | Active
--------|--------|-----------|-------
PWR_EN  |   0    |     0     |   1    
--------|--------|-----------|-------
RESETN  |   0    |     1     |   1    


We were putting the chip in reset mode when we really wanted standby. The main implication is that in reset mode the cable detection logic is turned off, so that's why my follow-up patches for using CABLE_DET as an interrupt failed.

I'll clean this up, rebase it (we already merged the voltage rail and PMIC stuff), and get a new patch series uploaded.

Patch-set: 6
Label: Code-Review=-2