bd99955: Don't monitor PROCHOT# to save power.
The BD99956 charger has a power save mode that it can enter once VBUS or
VCC is removed.  This commit changes the power save mode to leave BGATE
on, but not monitor the PROCHOT# signal.  It will re-enter its normal
mode once VBUS or VCC is present.

BUG=chrome-os-partner:55631
BRANCH=kevin
TEST=flash kevin; verify we can still charge; verify that power
consumption is less with this patch applied.

Change-Id: Ibab7ad30d5f1ae9917b46b40d6f2800ef19e52dd
Signed-off-by: Aseda Aboagye <aaboagye@google.com>