Update patch set 1
Patch Set 1: Code-Review-1

So, this is the reg write that will do what we want. However, I don't think this is the right way to do it. It's not guaranteed that all boards won't monitory PROCHOT#. Therefore, I'd like to have the board make that decision. I could introduce another CONFIG_* option?

Patch-set: 1
Label: Code-Review=-1