Cr50: Stub out the tpm_reset() function
In a future CL, we'll add the ability to reset just the TPM task
instead of the entire Cr50. This just implements a placeholder
function to call to do that.

BUG=chrome-os-partner:52366
BRANCH=none
TEST=make buildall; test on Cr50 hardware

Rebooting the AP triggers a SYS_RST_L pulse, as does running a
flashrom command over CCD:

  sudo flashrom -p raiden_debug_spi:target=AP --flash-name

Try both cases before & after this CL, observe the same behavior.

Change-Id: I4c4175dbc40a8ef1e1a5d2312c7df98a1ad5c756
Signed-off-by: Bill Richardson <wfrichar@chromium.org>