Update patch set 3
Patch Set 3:

Yes, that's the current state of things - Reef is 1:1 and other systems only have a single chip to control all ports.

The confusion is if we have a system where one chip controls multiple ports, but not all of them.

I'm new to this TCPC stuff so I don't know how likely that is. If it's unlikely then we don't need to complicate things further and we might want the port argument to be optional since it won't matter on systems where 1 TCPC/MCU/whatever controls multiple ports.

Patch-set: 3