Update patch set 2
Patch Set 2:

(2 comments)

The problem that this patch is solving is the case where the hw fifo is either not completely drained, or when the host reads more bytes without a write transaction preceding it. 

I am seeing cases of this during the verstage in Coreboot. It happens when the host is waiting for 'data_avail' bit to be set in the TPM STS register. At this time it's looping on reads of the TPM STS register. 

The i2c_read routine on the host side will first write the tpm address byte then after a fixed delay, attempt to read the number of bytes associated with a particular register. If this read fails for whatever reason, it will delay, then attempt the read again, without doing the register address write.

I think having the Cr50 I2CS always insert TPM register bytes where the HW read pointer is at the time the register address write completes is a safe thing to do. This improves the overall robustness so that the whole byte stream is not out of whack. All TPM fifo write/reads are preceded by reads of the STS register, so this is in my opinion a very safe place to adjust the fw read pointer, if necessary.

Patch-set: 2