Update patch set 2
Patch Set 2: Code-Review+1

This should be fine.  

INT_STATE is a peripheral control register, not part of the NVIC, so it doesn't affect the interrupt level that the core is running at.  

It can't cause preemption of the current interrupt handler because the current handler is running at the same priority as the next interrupt event.  If an event occurs during the handler, it will just cause the handler to be called again as soon as it exits.

Patch-set: 2
Reviewer: Gerrit User 1000187 <1000187@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1