Update patch set 6
Patch Set 6:

Renamed the function to more accurately describe its purpose. Also removed the check for buffered data during a TPM fifo register read. That's the only place the previous version of this function was being used. Now, it's only used for 1, 4 byte register reads and will always zero the i2cs read fifo depth. It still returns the depth when called, so that a calling function can flag an error or dump a log message if desired.

Patch-set: 6
Label: Code-Review=0