HACK: pwm: Revert internally to PWM resolution of 0-100
Commit 26c5a22 ("pwm: Increse PWM duty resolution to 16bits") broke the
backlight on Gru. The display switches off and on every few minutes, also
the actual brightness level does not change when the backlight is adjusted
through software. It is suspected that the issue might be related with
chrome-os-partner:56629.

As a stopgap solution revert the NPCX PWM code back to use internally a
resolution from 0-100. To avoid having to hack external tools we maintain
the external 16bit interface.

BRANCH=gru
BUG=chrome-os-partner:57260
TEST=build EC firmware for Gru and flash it; splash screen is shown
     during boot; the display brightness can be adjusted with the
     buttons.

Change-Id: I5fe95fcb58bad8ec0cf91412c02163f717f87c3e
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>